Refactor 2016 match insights to be more generic
