+++ android/src/main/java/com/thebluealliance/androidclient/subscribers/StatsListSubscriber.java
@@ -1,6 +1,5 @@
 package com.thebluealliance.androidclient.subscribers;
 
-import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
@@ -10,19 +9,16 @@ import com.thebluealliance.androidclient.comparators.StatListElementComparator;
 import com.thebluealliance.androidclient.database.Database;
 import com.thebluealliance.androidclient.eventbus.EventStatsEvent;
 import com.thebluealliance.androidclient.helpers.ThreadSafeFormatters;
-import com.thebluealliance.androidclient.listitems.EventTypeHeader;
-import com.thebluealliance.androidclient.listitems.LabelValueListItem;
 import com.thebluealliance.androidclient.listitems.ListItem;
 import com.thebluealliance.androidclient.listitems.StatsListElement;
 import com.thebluealliance.androidclient.models.Team;
+import com.thebluealliance.androidclient.renderers.insights.EventInsights2016Renderer;
+import com.thebluealliance.androidclient.renderers.insights.EventInsightsRenderer;
 
 import org.greenrobot.eventbus.EventBus;
 
 import android.content.res.Resources;
-import android.support.annotation.Nullable;
-import android.support.annotation.StringRes;
 
-import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -30,7 +26,6 @@ import java.util.Map.Entry;
 
 public class StatsListSubscriber extends BaseAPISubscriber<StatsListSubscriber.Model, List<ListItem>> {
 
-    private static DecimalFormat df = new DecimalFormat("#.##");
 
     private String mStatToSortBy;
     private Resources mResources;
@@ -100,8 +95,15 @@ public class StatsListSubscriber extends BaseAPISubscriber<StatsListSubscriber.M
         Collections.sort(mTeamStats, new StatListElementComparator(mStatToSortBy));
 
         // Event stats
-        if (mEventYear == 2016 && mAPIData.getInsights() != null && mAPIData.getInsights().isJsonObject()) {
-            generateEventInsights(mAPIData.getInsights().getAsJsonObject());
+        EventInsightsRenderer insightsRenderer = null;
+        switch (mEventYear) {
+            case 2016:
+                insightsRenderer = new EventInsights2016Renderer(mEventStats, mResources);
+                break;
+        }
+
+        if (insightsRenderer != null) {
+            insightsRenderer.generateEventInsights(mAPIData.getInsights());
         }
 
         mEventBus.post(new EventStatsEvent(getTopStatsString()));
@@ -110,134 +112,6 @@ public class StatsListSubscriber extends BaseAPISubscriber<StatsListSubscriber.M
     @Override
     public boolean isDataValid() {
         return super.isDataValid() && mAPIData.getStats() != null && mAPIData.getStats().isJsonObject();
-    }
-
-    private void generateEventInsights(JsonObject eventInsights) {
-        mEventStats.clear();
-        JsonObject qualData;
-        JsonObject elimData;
-        if (eventInsights.has("qual") && eventInsights.get("qual").isJsonObject()) {
-            qualData = eventInsights.get("qual").getAsJsonObject();
-        } else {
-            qualData = new JsonObject();
-        }
-        if (eventInsights.has("playoff") && eventInsights.get("playoff").isJsonObject()) {
-            elimData = eventInsights.get("playoff").getAsJsonObject();
-        } else {
-            elimData = new JsonObject();
-        }
-
-        generate2016MatchInsights(qualData, elimData);
-        generate2016DefenseInsights(qualData, elimData);
-        generate2016TowerInsights(qualData, elimData);
-    }
-
-    private @StringRes int[] matchTitles = {R.string.breakdown2016_avg_low_goal, R.string
-            .breakdown2016_avg_high_goal, R.string.breakdown2016_avg_score, R.string
-            .breakdown2016_avg_win_score, R.string.breakdown2016_avg_win_margin, R.string
-            .breakdown2016_avg_auto_score, R.string.breakdown2016_teleop_cross, R.string
-            .breakdown2016_avg_boulder_score, R.string.breakdown2016_avg_tower_score, R.string.breakdown2016_avg_foul_score};
-    private String[] matchKeys = {"average_low_goals", "average_high_goals", "average_score",
-            "average_win_score", "average_win_margin", "average_auto_score",
-            "average_crossing_score", "average_boulder_score", "average_tower_score", "average_foul_score"};
-
-    private void generate2016MatchInsights(JsonObject quals, JsonObject elims) {
-        mEventStats.add(new EventTypeHeader(mResources.getString(R.string.breakdown_match_stats)));
-
-        String qualFormat = mResources.getString(R.string.breakdown2016_qual);
-        String elimFormat = mResources.getString(R.string.breakdown2016_elim);
-
-        // high_score
-        String qualHighScore = null, elimHighScore = null;
-        if (quals.has("high_score") && quals.get("high_score").isJsonArray()) {
-            JsonArray qualHigh = quals.get("high_score").getAsJsonArray();
-            qualHighScore = mResources.getString(R.string.breakdown2016_match_stat, qualHigh
-                    .get(0).getAsInt(), qualHigh.get(2).getAsString());
-        }
-        if (elims.has("high_score") && elims.get("high_score").isJsonArray()) {
-            JsonArray elimHigh = elims.get("high_score").getAsJsonArray();
-            elimHighScore = mResources.getString(R.string.breakdown2016_match_stat, elimHigh
-                    .get(0).getAsInt(), elimHigh.get(2).getAsString());
-        }
-        mEventStats.add(new LabelValueListItem(mResources.getString(R.string
-                .breakdown2016_high_score),
-                combineQualAndElimStat(qualHighScore, elimHighScore),
-                true));
-
-        for (int i = 0; i < matchKeys.length; i++) {
-            String qualStat = null, elimStat = null;
-            if (quals.has(matchKeys[i]) && quals.get(matchKeys[i]).isJsonPrimitive()) {
-                qualStat = df.format(quals.get(matchKeys[i]).getAsDouble());
-            }
-            if (elims.has(matchKeys[i]) && elims.get(matchKeys[i]).isJsonPrimitive()) {
-                elimStat = df.format(elims.get(matchKeys[i]).getAsDouble());
-            }
-            mEventStats.add(new LabelValueListItem(mResources.getString(matchTitles[i]),
-                    combineQualAndElimStat(qualStat, elimStat), true));
-        }
-    }
-
-    private String[] defenseTitle = {"LowBar", "A_ChevalDeFrise", "A_Portcullis", "B_Ramparts",
-            "B_Moat", "C_SallyPort", "C_Drawbridge", "D_RoughTerrain", "D_RockWall", "breaches"};
-    private @StringRes int[] defenseName = {R.string.defense2016_low_bar, R.string
-            .defense2016_cdf, R.string.defense2016_portcullis, R.string.defense2016_ramparts, R
-            .string.defense2016_moat, R.string.defense2016_sally_port, R.string
-            .defense2016_drawbridge, R.string.defense2016_rough_terrain, R.string
-            .defense2016_rock_wall, R.string.defense2016_breaches};
-
-    private void generate2016DefenseInsights(JsonObject quals, JsonObject elims) {
-        mEventStats.add(new EventTypeHeader(mResources.getString(R.string.breakdown2016_defense_stats)));
-        String defenseFormat = mResources.getString(R.string.defense2016_percent_format);
-        for (int i = 0; i < defenseName.length; i++) {
-            String qualStat = null, elimStat = null;
-            if (quals.has(defenseTitle[i]) && quals.get(defenseTitle[i]).isJsonArray()) {
-                JsonArray qualData = quals.get(defenseTitle[i]).getAsJsonArray();
-                qualStat = String.format(defenseFormat, qualData.get(0).getAsInt(), qualData.get(1).getAsInt(),
-                        qualData.get(2).getAsDouble());
-            }
-            if (elims.has(defenseTitle[i]) && elims.get(defenseTitle[i]).isJsonArray()) {
-                JsonArray elimData = elims.get(defenseTitle[i]).getAsJsonArray();
-                elimStat = String.format(defenseFormat, elimData.get(0).getAsInt(), elimData.get(1)
-                        .getAsInt(), elimData.get(2).getAsDouble());
-            }
-            mEventStats.add(new LabelValueListItem(mResources.getString(defenseName[i]),
-                    combineQualAndElimStat(qualStat, elimStat), true));
-        }
-    }
-
-    private @StringRes int[] towerTitles = {R.string.breakdown2016_challenge, R.string
-            .breakdown2016_scales, R.string.breakdown2016_captures};
-    private String[] towerKeys = {"challenges", "scales", "captures"};
-
-    private void generate2016TowerInsights(JsonObject quals, JsonObject elims) {
-        mEventStats.add(new EventTypeHeader(mResources.getString(R.string.breakdown2016_tower_stats)));
-        String defenseFormat = mResources.getString(R.string.defense2016_percent_format
-        );
-        for (int i = 0; i < towerTitles.length; i++) {
-            String qualStat = null, elimStat = null;
-            if (quals.has(towerKeys[i]) && quals.get(towerKeys[i]).isJsonArray()) {
-                JsonArray qualData = quals.get(towerKeys[i]).getAsJsonArray();
-                qualStat = String.format(defenseFormat, qualData.get(0).getAsInt(), qualData.get(1)
-                        .getAsInt(), qualData.get(2).getAsDouble());
-            }
-            if (elims.has(towerKeys[i]) && elims.get(towerKeys[i]).isJsonArray()) {
-                JsonArray elimData = elims.get(towerKeys[i]).getAsJsonArray();
-                elimStat = String.format(defenseFormat, elimData.get(0).getAsInt(), elimData.get(1)
-                        .getAsInt(), elimData.get(2).getAsDouble());
-            }
-            mEventStats.add(new LabelValueListItem(mResources.getString(towerTitles[i]),
-                    combineQualAndElimStat(qualStat, elimStat), true));
-        }
-    }
-
-    private String combineQualAndElimStat(@Nullable String qualStat, @Nullable String elimStat) {
-        if (qualStat != null && elimStat != null) {
-            return mResources.getString(R.string.breakdown2016_qual_and_elim, qualStat, elimStat);
-        } else if (qualStat != null) {
-            return mResources.getString(R.string.breakdown2016_qual, qualStat);
-        } else {
-            return mResources.getString(R.string.breakdown2016_elim, elimStat);
-        }
     }
 
     private String getTopStatsString() {
